履职工作日志查询接口和查询接口导出结果bug修复

+++ schema/report_schema.py
@@ -15,7 +15,7 @@ class DailyWorkResponse(BaseModel):
     title: str
     agenda: Optional[str] = None
     text_message: Optional[str] = None
-    duration_minutes: Optional[int] = None
+    duration_minutes: Optional[float] = None
     participant_names: List[str] = []
     company_names: List[str] = []
